# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sisCotacao — API",
    "description": "Contrato dos endpoints principais do ERP (fase 1). Autenticação por Bearer token emitido em /api/login (TTL 7 dias, revogável por logout ou troca de senha). Portal do fornecedor usa token próprio na URL e está fora deste contrato.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem (nginx proxy /api)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/pronto": {
      "get": {
        "tags": [
          "infra"
        ],
        "summary": "Readiness real — executa SELECT 1 no banco",
        "security": [],
        "responses": {
          "200": {
            "description": "Sistema utilizável",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProntoOK"
                }
              }
            }
          },
          "503": {
            "description": "Banco indisponível",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Autentica e emite o token da sessão",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOut"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas ou usuário inativo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sistema/status": {
      "get": {
        "tags": [
          "sistema"
        ],
        "summary": "Versão app × schema e pendências por risco",
        "responses": {
          "200": {
            "description": "Estado de atualização",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SistemaStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutenticado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            },
            "description": "Erro de infraestrutura"
          }
        }
      }
    },
    "/api/sistema/updates/apply": {
      "post": {
        "tags": [
          "sistema"
        ],
        "summary": "Aplica migrações pendentes por nível de risco",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Migrações aplicadas (ou nada pendente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SistemaStatus"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao aplicar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/sistema/updates/log": {
      "get": {
        "tags": [
          "sistema"
        ],
        "summary": "Histórico de atualizações aplicadas (últimos 50)",
        "responses": {
          "200": {
            "description": "Eventos mais recentes primeiro",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "log": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AtualizacaoLog"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            },
            "description": "Erro de infraestrutura"
          }
        }
      }
    },
...
    "/api/perfis": {
      "get": {
        "tags": [
          "permissoes"
        ],
        "summary": "Perfis com matriz de permissões por recurso",
        "responses": {
          "200": {
            "description": "Perfis (Administrador marcado como superuser)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerfilAcesso"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "permissoes"
        ],
        "summary": "Cria um perfil novo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilCRUDIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Perfil criado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nome reservado/duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clientes": {
      "get": {
        "tags": [
          "clientes"
        ],
        "summary": "Lista clientes (com segmento/categoria)",
        "parameters": [
          {
            "name": "somente_ativos",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "vendedor_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clientes"
        ],
        "summary": "Cria cliente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fornecedores": {
      "get": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Lista fornecedores (busca por termo/categoria)",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoria",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "somente_ativos",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de fornecedores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fornecedor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fornecedores"
        ],
        "summary": "Cria fornecedor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FornecedorIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fornecedor criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plano-contas": {
      "get": {
        "tags": ["financeiro"],
        "summary": "Lista o plano de contas com filtros de natureza e rateabilidade",
        "parameters": [
          {"name": "tipo", "in": "query", "schema": {"type": "string", "enum": ["receita", "despesa"]}},
          {"name": "natureza", "in": "query", "schema": {"type": "string"}},
          {"name": "rateavel", "in": "query", "schema": {"type": "boolean"}},
          {"name": "somente_ativos", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {"200": {"description": "Contas do plano"}, "400": {"$ref": "#/components/responses/ErroValidacao"}}
      },
      "post": {
        "tags": ["financeiro"],
        "summary": "Cria conta de receita ou despesa com classificação",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PlanoContaIn"}}}},
        "responses": {"201": {"description": "Conta criada"}, "400": {"$ref": "#/components/responses/ErroValidacao"}}
      }
    },
...
    "/api/financeiro/competencias": {
      "get": {
        "tags": ["financeiro"],
        "summary": "Lista competências financeiras para precificação",
        "responses": {"200": {"description": "Competências cadastradas"}}
      },
      "post": {
        "tags": ["financeiro"],
        "summary": "Cria ou atualiza uma competência financeira",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CompetenciaFinanceiraIn"}}}
        },
        "responses": {"201": {"description": "Competência criada"}, "400": {"$ref": "#/components/responses/ErroValidacao"}}
      }
    },
...
    "/api/webhooks/logs": {
      "get": {
        "tags": [
          "financeiro"
        ],
        "summary": "Logs de webhooks de pagamento (auditoria)",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "processado|duplicado|ignorado|nao_autorizado|erro|payload_invalido|caixa_pendente"
          },
          {
            "name": "desde",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de logs"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "HMAC (base64.payload.signature), TTL 7 dias"
      }
    },
...
    "schemas": {
      "Erro": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "db_indisponivel"
          }
        }
      },
      "ProntoOK": {
        "type": "object",
        "properties": {
          "pronto": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "LoginIn": {
        "type": "object",
        "required": [
          "login",
          "senha"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "senha": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "LoginOut": {
        "type": "object",
        "properties": {
          "autenticado": {
            "type": "boolean",
            "example": true
          },
          "token": {
            "type": "string",
            "description": "Bearer para as demais rotas"
          },
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "admin",
              "vendedor"
            ]
          }
        }
      },
...
    }
  }
}